requirements.txt
aiogram~=3.23.0
python-dotenv~=1.2.1
aiosqlite~=0.21.0